+++ packages/vue/src/components/focus-trap/index.ts
@@ -1,0 +1,2 @@
+export { default as FocusTrap } from './focus-trap.vue'
+export type { FocusTrapBaseProps, FocusTrapProps } from './focus-trap.vue'